Switch the README to `let` and arrow functions
Users who are writing code nowadays will be using the LTS 6.x or
the latest 7.x releases. An up-to-date installation will support
ES6 features such as `let` and arrow functions. There is no negative
impact in providing all the example code in the README as such,
even more so since parts of the example code there already utilize
arrow functions, such as for promise callbacks.