adds usage example for the @sendgrid/helpers EmailAddress class
## Description

I recently wanted to make use of the `EmailAddress` in the `@sendgrid/helpers` package but found that there was not usage example in the documentation. I had to read the code to figure out a way to use it.

Hence this PR, Will add more usage for other helper classes like `Personalization` etc. 

## Why raising the PR

- In case there is someone else who is just like me search the entire doc for an example of where the `Helper` classes are being used.